login and signup page metadata
